% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crisis_periods}
\alias{crisis_periods}
\title{Crisis Sub-Period Definitions}
\format{
A named list with eight elements:
\describe{
  \item{PreCrisis}{Pre-Crisis Baseline (12 Jan 2006 - 31 Jul 2007)}
  \item{GFC}{Global Financial Crisis (1 Aug 2007 - 30 Jun 2009)}
  \item{ESDC}{European Sovereign Debt Crisis (1 Dec 2009 - 30 Jun 2012)}
  \item{CSC}{Chinese Stock Crash (15 Jun 2015 - 31 Dec 2016)}
  \item{PreCOVID}{Pre-COVID interval (1 Jan 2017 - 31 Jan 2020)}
  \item{COVID}{COVID-19 Pandemic (1 Feb 2020 - 31 Dec 2021)}
  \item{RusUkr}{Russia-Ukraine episode (1 Feb 2022 - 31 Dec 2023)}
  \item{MidEastTariffs}{Middle-East tensions and tariffs (1 Jan 2024 - 18 Mar 2026)}
}
}
\usage{
crisis_periods
}
\description{
A named list of length-two character vectors specifying the start and
end dates of the eight crisis sub-periods analysed in the paper.
}
\examples{
data(crisis_periods)
crisis_periods$GFC
}
\keyword{datasets}
