% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{walktrap_communities}
\alias{walktrap_communities}
\title{Walktrap Community Detection on a Contagion Network}
\usage{
walktrap_communities(g)
}
\arguments{
\item{g}{An igraph object.}
}
\value{
An integer vector of community memberships, one per vertex; or
  \code{NULL} if the network has too few edges.
}
\description{
Detects communities using the Walktrap algorithm of Pons and Latapy
(2006) on the symmetrised version of the directed contagion network.
}
\examples{
m <- matrix(runif(25, 0, 0.1), 5, 5); diag(m) <- 0
g <- build_network(m, threshold = 0.02)
walktrap_communities(g)
}
